scripts/requirements-scrape.txt
playwright==1.49.0
google-cloud-storage>=2.0.0
